Import scipy.ndimage before tensorflow to fix jpeg load. (#165)
When importing tensorflow before scipi.ndimage and running the
example in a docker container, the following error is given:

```
 % python example.py
Traceback (most recent call last):
  File "example.py", line 25, in <module>
    im = im / 255.
TypeError: unsupported operand type(s) for /: 'instance' and 'float'
```
This only happens for 'cat.jpg'. When converting the image to a '.png'
file, the example runs as expected.

When swapping the imports around so that scipy.ndimage is imported
BEFORE tensorflow, both 'jpg' and 'png' files work.
Former-commit-id: c716c1b8b55d0e318d6e8a8f214bc6ecba382a45
Former-commit-id: 1d08722e84cb23fbc09293b1649575730f4c8558