Import scipy.ndimage before tensorflow to fix jpeg load. (#165)
When importing tensorflow before scipi.ndimage and running the
example in a docker container, the following error is given:

```
 % python example.py
Traceback (most recent call last):
  File "example.py", line 25, in <module>
    im = im / 255.
TypeError: unsupported operand type(s) for /: 'instance' and 'float'
```
This only happens for 'cat.jpg'. When converting the image to a '.png'
file, the example runs as expected.

When swapping the imports around so that scipy.ndimage is imported
BEFORE tensorflow, both 'jpg' and 'png' files work.
Former-commit-id: 28c687a8f4d2bd909fc60ac04effce6be6e992a5 [formerly caf0fed0c5694bc0d8adebbb055cbb62eaeb8b26] [formerly 1f27437150ba3a570b58bde82499fc426801a440] [formerly 85ccf9f0948f1766807c0e3a8adb6d18a1912883 [formerly 1d08722e84cb23fbc09293b1649575730f4c8558] [formerly c716c1b8b55d0e318d6e8a8f214bc6ecba382a45]]
Former-commit-id: 85ccf9f0948f1766807c0e3a8adb6d18a1912883 [formerly 1d08722e84cb23fbc09293b1649575730f4c8558]
Former-commit-id: 16e106145e0d521b0f5cc09b491fad48e3f2062d
Former-commit-id: 91e8ec46d6e6a97eaa14c61bc1085ef50c9d1e09