Wrap the cifar10 multigpu model construction part with a variable_scope
Without the new variable_scope, creating apply_gradient_op raises
an error that additional moving average or slot variables could not
be created. This is because of the 'leaky reuse' of variable scope,
so we correct the problem by explicitly introducing a new variable scope.

Related issues: tensorflow/models#901, tensorflow/tensorflow#6220


Former-commit-id: 9d96e9fe2b40f0da266261e2200cfc3c27f97b87
Former-commit-id: 3e6006ea32b919459935b118f8d50eb0c5c9dba2