Wrap the cifar10 multigpu model construction part with a variable_scope
Without the new variable_scope, creating apply_gradient_op raises
an error that additional moving average or slot variables could not
be created. This is because of the 'leaky reuse' of variable scope,
so we correct the problem by explicitly introducing a new variable scope.

Related issues: tensorflow/models#901, tensorflow/tensorflow#6220


Former-commit-id: d7c620ef39b435bcae3feadc7d58c0d07e4577e2 [formerly 6f8cf6c7c7c80c6e71b16f49a69623884fd7ea09 [formerly 3e6006ea32b919459935b118f8d50eb0c5c9dba2] [formerly 9d96e9fe2b40f0da266261e2200cfc3c27f97b87]]
Former-commit-id: 072f46e1090283c4632e5fc6912373e1b48f150d [formerly 34cbf4c5230ead9f4fbccf0cee78baa9e6f73ce0] [formerly ad0ba60fc0684cc685ea229368ed196e6773f589]
Former-commit-id: c37da7b94d22f94f7612a1b2f3e9bde93593a8de
Former-commit-id: 7bec9ba24787599676490faefb467d88b515b859