Add session hook for benchmark metric logging. (#3672)
* Add session hook for benchmark metric logging.

Current hook is very similar as the LoggingTensorHook. Some of the
function are directly copied since the original one was not
exposed for import. We should seek to eventually move this code to
core when it is mature enough.

* Update metric_hook to use LoggingTensorHook as base.

The existing hook is similar enough to LoggingTensorHook, and
we should eliminate duplicate as much as possible.

* Address review comment.

1. Update global step tensor handle.
2. Update tests.
3. Update document.

* Update tests for py3.

* Fix lint error


Former-commit-id: 5d57c1e745faf2e05302261b542cec5b7a751b70 [formerly e3520eb86beeef3350ef7224d9c9c18c3e45a550] [formerly 4b85dab1e9d3a4c541d873cb473ee2e82dd810c9]
Former-commit-id: 5e8868679e264de3bca3a7a0fe960d3d86699e70
Former-commit-id: ee1de2085c4a4a1d57b353ab7d183d81fbb7bbc2