mnist_eager to use GPU if not specified otherwise (#4216)
Update mnist_eager.py to use GPU if available and not specified otherwise. At the moment it uses the CPU even if a GPU is available, since "if flags_obj.no_gpu or tf.test.is_gpu_available():" in "def run_mnist_eager(flags_obj):" evaluates to True if a GPU is available.

Former-commit-id: db3a00922368ecd297de2844e7cb179c4211a291 [formerly 160c7b34870e0741a9b3a0ad49fa48f7c2ec5e72 [formerly 9b2830ea4c488724fd764aa4d9a32390c699e1bf] [formerly d0b6a34bb0dbd981e3785661f04f04cde9c4222b]]
Former-commit-id: a8d8bb7cc76423d1528e93bcbbb3d7c70bbfd44a [formerly 8758c962eda7bfa3064a806e276f9de1eb1b818f] [formerly f834c28db94253f26ca4f6bea45b2d08a1526c9d]
Former-commit-id: edc679a5841f50beff624efd407ed9193ae5663f
Former-commit-id: 07723b59b3fd8cffdd393677233d7917a94e14de