mnist_eager to use GPU if not specified otherwise (#4216)
Update mnist_eager.py to use GPU if available and not specified otherwise. At the moment it uses the CPU even if a GPU is available, since "if flags_obj.no_gpu or tf.test.is_gpu_available():" in "def run_mnist_eager(flags_obj):" evaluates to True if a GPU is available.

Former-commit-id: 160c7b34870e0741a9b3a0ad49fa48f7c2ec5e72 [formerly 9b2830ea4c488724fd764aa4d9a32390c699e1bf] [formerly d0b6a34bb0dbd981e3785661f04f04cde9c4222b]
Former-commit-id: f834c28db94253f26ca4f6bea45b2d08a1526c9d
Former-commit-id: 8758c962eda7bfa3064a806e276f9de1eb1b818f