Bug fix: change dict.iteritems() to dict.items()
`iteritems()` was removed from python3. `items()` does the same functionality so changing it will work in both python2 and python3. The only difference as far as I know is `iteritems()` returns a generator where `items` returns a list. But for this this code it will not make any difference where we are just changing the key of the dict to a string.

Former-commit-id: 72e1f97550c11b8bde75158a2e24fa9ae013d60d [formerly e4fd3c7138fa8c69fb119f6f3259a94ce0824102 [formerly f2e13b0551944ceb64bf529552f425b9ef7a0f17] [formerly e2ea9eb4e2ab0fecc9adfabbec838dbfd114be29]]
Former-commit-id: d397224f38ade39ded4edbe55ba09c9e6db01968 [formerly d45aef52a66c397c638495a71bcd6389fcb5032a] [formerly cc3d005b24896f3aba7487f486110ab60855dbe0]
Former-commit-id: 818d85b09688f341e0d6bebcc699bf21c5e2a13b
Former-commit-id: f2719d8dd4cf1d2dd449d17451fb873274ae106d