Fix bug where data_async_generation.py would freeze. (#4989)
The data_async_generation.py process would print to stderr, but the main process would redirect it's stderr to a pipe. The main process never read from the pipe, so when the pipe was full, data_async_generation.py would stall on a write to stderr. This change makes data_async_generation.py not write to stdout/stderr.

Former-commit-id: c7c58ddc2a906a2b3ba1d2f0820a084b2968f0b4 [formerly 49a9bd31b0f1fd3d32067eb14a9fe387eed59a79] [formerly d452b878b298e2a0827be8d995269f6b859bd70d] [formerly 6543d63e678edcd696e03a5eaf228b7a615aa978 [formerly aad1b43d49821c8268b0e32a890d60961051d3c7] [formerly 58037d2cbfe24fc3d1db9fc74caa396db6475ea5]]
Former-commit-id: 6543d63e678edcd696e03a5eaf228b7a615aa978 [formerly aad1b43d49821c8268b0e32a890d60961051d3c7]
Former-commit-id: 0e75f67c18f1e41152683d7b9cb81bc64bc02946
Former-commit-id: b963b1c9890603dd2d4a52f20544196c35c89971