Update the runtime version in bash script
The bash script to submit training job for pets detection has runtime-version of 1.8. This will trigger `TypeError: non_max_suppression() got an unexpected keyword argument 'score_threshold'` on the Google Cloud since 1.8 and older does not support this keyword argument. Therefore, update this runtime version to 1.9, which is the most recent runtime version that is published on June 27, 2018. 

See 
https://github.com/tensorflow/models/issues/5056
https://cloud.google.com/ml-engine/docs/tensorflow/runtime-version-list
Former-commit-id: f134ad7b92831e9afd5bc9a90bbae8af2d59a2aa [formerly 2bfef36d5878c7a0a02f28a326d739426069b5f1 [formerly ed7aae1c9bfbf9927eb66bb5835dbb03245bf8ff] [formerly fc918584cf35b093928a96916c503c9a17de1633]]
Former-commit-id: dfdfce4b2c8323844c6da087346debaa4fccd501 [formerly b33c2f094e6292b3a8ee1455ba88caf37308ab9f] [formerly ea74f86aef32eb88024e50d680acd1907c74c066]
Former-commit-id: a1d114f93ff3a2bf6d41ad98e89eb3b868e39742
Former-commit-id: 9b72752967ffbb089e5b4107056c626a1597adc3