Add `--image_bytes_as_serving_input` flag to export SavedModel (#5393)
with serving signature that accepts JPEG image bytes instead
of a fixed size [HxWxC] image tensor.

Passing JPEG image bytes is easier for inference/serving use
cases. The model internally resizes/crops the JPEG image to
required [HxWxC] tensor before passing it on for actual model
inference.

This change aligns with Cloud TPU/ResNet-50 model that offers a
similar interface (jpeg bytes) for inferencing here:

https://github.com/tensorflow/tpu/tree/master/models/official/resnet

NOTE: This flag is set to `True` by default for ImageNet, and is
disallowed for CIFAR (as it does not apply to CIFAR).
Former-commit-id: cf3c58aca3104de6f8c707023363b71986792d36 [formerly c092ed3f9d86663d5a1bb54f0cf374ac83839129] [formerly e9cbe3773a0607300fa5833db4ea1b3d0e30d9f9] [formerly abe4de39bc56d73d858a2147b960eb12b06d67b5 [formerly dc06bbb5f327fe565d9eb7c3c3b79d60a63d28bd] [formerly 3578c2e4fb18d12aabfbbe2240fddc7eca07cde1]] [formerly d7610c63d56888a1aa687f016c59d0e46504dada [formerly 0a67e625cbb802ec8561650c693d51aaf51d2219 [formerly db428e440f63153b42f87c1f6ea8bf8ae143f865] [formerly 295259c324732e2cbc70e32e74ac1cd61a125c28]]]
Former-commit-id: 0a67e625cbb802ec8561650c693d51aaf51d2219 [formerly db428e440f63153b42f87c1f6ea8bf8ae143f865]
Former-commit-id: 6963f2ef354513ee64ee800547f82db4c32b8556 [formerly b4ab6dc2316a0bac689bfb688615f36f845edbda] [formerly 0c64db8355b269b4416c6ccec027880b616ce95d]
Former-commit-id: fea3e4185ac4e6cb7d4eb56c90e90d5d4fffe228
Former-commit-id: 89220c4843e8b7fff1e831419798946d17441001