Add `--image_bytes_as_serving_input` flag to export SavedModel (#5393)
with serving signature that accepts JPEG image bytes instead
of a fixed size [HxWxC] image tensor.

Passing JPEG image bytes is easier for inference/serving use
cases. The model internally resizes/crops the JPEG image to
required [HxWxC] tensor before passing it on for actual model
inference.

This change aligns with Cloud TPU/ResNet-50 model that offers a
similar interface (jpeg bytes) for inferencing here:

https://github.com/tensorflow/tpu/tree/master/models/official/resnet

NOTE: This flag is set to `True` by default for ImageNet, and is
disallowed for CIFAR (as it does not apply to CIFAR).
Former-commit-id: 0a67e625cbb802ec8561650c693d51aaf51d2219 [formerly db428e440f63153b42f87c1f6ea8bf8ae143f865] [formerly 295259c324732e2cbc70e32e74ac1cd61a125c28]
Former-commit-id: 3578c2e4fb18d12aabfbbe2240fddc7eca07cde1
Former-commit-id: dc06bbb5f327fe565d9eb7c3c3b79d60a63d28bd