Add distribute strategies to transformer. (#6883)
* Fixes that make transformer run.

* Remove debug print statements.

* Changed the permissions to 644.

* Fix the rest of the permissions.

* enable static batch in all benchmarks

* Restrict dist strat hack to training mode

For now we will do predict/eval without dist strat, so remove that hack in non training cases.

* Use `inputs` instead of `x` as arg name for call

Keras has different behavior based on whether the inputs are called `inputs` or not. Using `inputs` gives expected behaviors.

* Avoid extra map fn on input in dist strat case

* Update how we handle custom metrics

This new approach works with and without dist strat. The previous one didn't work with dist strat. We need to fix that but this is reasonable in meantime (b/133724664).

* Update benchmarks

* typo in metrics code

* Revert metrics change

Didn't actually work in distributed case..


Former-commit-id: 15df1eed406555fdb933c52d222cb9ae4432a5a2 [formerly 6ce63fa8b518f146b214ec8a29ced67d6f2d6d8e] [formerly b9c1d1ca228819c212d0f1b227bebd3a0fb3e263]
Former-commit-id: 95425483546f93319d714db69bfe4da7e1746b29
Former-commit-id: 24f1e51159a2c5c231877daa57f22063de1daf0e