Simply LayerNorm mixed precision logic.
Instead of needing to ensure variables are float32, casting inputs to float32, etc, instead dtype="float32" is passed to the layer constructor, which will do all that logic automatically.

The only difference is the output of LayerNorm is now float32 instead of float16, so an extra cast is needed elsewhere.

PiperOrigin-RevId: 273833286

Former-commit-id: 0257b276a10b12f2c414fd5526f4103ffe24d7c9
Former-commit-id: 3d5183a09fcce0f483b379563d272bb42747dcac