Simply LayerNorm mixed precision logic.
Instead of needing to ensure variables are float32, casting inputs to float32, etc, instead dtype="float32" is passed to the layer constructor, which will do all that logic automatically.

The only difference is the output of LayerNorm is now float32 instead of float16, so an extra cast is needed elsewhere.

PiperOrigin-RevId: 273833286

Former-commit-id: f9c724c7d369aefec8adc456bba3663d0f4169ca [formerly c663023f10b298bde365cfdfe2a369659504d181 [formerly 3d5183a09fcce0f483b379563d272bb42747dcac] [formerly 0257b276a10b12f2c414fd5526f4103ffe24d7c9]]
Former-commit-id: 64d70fe159d0cb88e3864f2a2045bba6c9e4adb5 [formerly dfb80e85d6fa19f89c2d62fd5b50e604bf883bfa] [formerly 4b48bf686d48b432248e53658049c865421b2408]
Former-commit-id: 689b76152d5deebcde97ecafd52fe3a666a02686
Former-commit-id: 64a5dae6e536b2eb9ef954c9cb294282c3c303d7