Temporarily disable explicit allreduce in BERT SQuAD
In BERT SQuAD, disable explicit allreduce for now to keep the original clip_by_global_norm math. With explicit allreduce, the gradients before allreduce are scaled so even if we move clip_by_global_norm before allreduce (as in TF1 and pre-TF 2.2) it will operate on scaled gradients, the math will be changed. So with explicit allreduce, it is better to move clip_by_global_norm to after allreduce.

PiperOrigin-RevId: 299278082

Former-commit-id: 11ccb99eae06471822a80beb4ba3f65faedd0fe8
Former-commit-id: a53ee01cb740bb39f7096422a398d7f996d5becc
Former-commit-id: cb7e5c4d930d70b3b058cc8a92449f78e7074329