Temporarily disable explicit allreduce in BERT SQuAD
In BERT SQuAD, disable explicit allreduce for now to keep the original clip_by_global_norm math. With explicit allreduce, the gradients before allreduce are scaled so even if we move clip_by_global_norm before allreduce (as in TF1 and pre-TF 2.2) it will operate on scaled gradients, the math will be changed. So with explicit allreduce, it is better to move clip_by_global_norm to after allreduce.

PiperOrigin-RevId: 299278082

Former-commit-id: 2150d348fd1a03c6d9d70d836ae8c5e618b5dafd [formerly 11ccb99eae06471822a80beb4ba3f65faedd0fe8]
Former-commit-id: 14a16d516b74bd13f753c979c14ad096c4ce6f8f [formerly cb7e5c4d930d70b3b058cc8a92449f78e7074329] [formerly a53ee01cb740bb39f7096422a398d7f996d5becc]
Former-commit-id: b3829b6e559039aa5efb1e5c4ced0369be3c552b
Former-commit-id: 1e96d9e015a8aba771aa8f42d6572e1939c56c60