Save to tmp directory on non-chief workers in model_training_utils
In a multi worker set up saving is done on each worker. If they're saving to the same location, e.g. GCS, there will be conflicts. With this change we save to temporary directory on non-chief workers.

Note that, there may be synchronization in saving that needs all workers to participate, so we cannot only save on one worker.

PiperOrigin-RevId: 300141152

Former-commit-id: eed25789147ce08dc00adc3320b2bf55b72d897c [formerly 48fd23912096c4eb2b072a38f854c1d37958076c] [formerly 70abe176e15b4482e03ab991cedd87f9f8313f3f] [formerly 34f3f67c94bb447a931aea627b4dc4a5bd16bcad [formerly cac305799b8c18ecc830c61a258921dcfe0213d5] [formerly 0c368204ca53038beb9a7fd814b8d4d0cbbb0cf6]] [formerly 0ce4fb0184874f3efe2b673b2a67270b6c0bd60e [formerly 682d36b5c7c88ef3eaa4322b0af9727d40e41d55]]
Former-commit-id: c9ef59aec0427b619315541068bb2bef7d80813e
Former-commit-id: a56c3b2d33ea97a628c9a9dc0c3ed79fdb01234d
Former-commit-id: d2336381834953c8498fe53a072120ac7f60e3ff
Former-commit-id: 7782b02434c8f65d2c1772e1c8dcd1ba4120cdba