Save to tmp directory on non-chief workers in model_training_utils
In a multi worker set up saving is done on each worker. If they're saving to the same location, e.g. GCS, there will be conflicts. With this change we save to temporary directory on non-chief workers.

Note that, there may be synchronization in saving that needs all workers to participate, so we cannot only save on one worker.

PiperOrigin-RevId: 300141152

Former-commit-id: 0ce4fb0184874f3efe2b673b2a67270b6c0bd60e [formerly 682d36b5c7c88ef3eaa4322b0af9727d40e41d55]
Former-commit-id: 34f3f67c94bb447a931aea627b4dc4a5bd16bcad [formerly cac305799b8c18ecc830c61a258921dcfe0213d5] [formerly 0c368204ca53038beb9a7fd814b8d4d0cbbb0cf6]
Former-commit-id: 70abe176e15b4482e03ab991cedd87f9f8313f3f
Former-commit-id: 48fd23912096c4eb2b072a38f854c1d37958076c