Use float32 activation in Transformer.
Float32 is used if the model uses mixed precision with bfloat16. Float16 activation are unchanged.

The motivation is that BERT with the LAMB optimizer with a gelu activation has an unstable loss when gelu is in bfloat16. Unfortunately, it is not easy to check if the LAMB optimizer and gelu is used, and perhaps there are other cases that work better with float32 activations instead of bfloat16 activations, so we always do the activation in float32 instead of bfloat16.

PiperOrigin-RevId: 313618322

Former-commit-id: d25d92c54bad3a615a2e4e78d0a642e71712aeb1 [formerly 10fbc703b25a60bb053c60c81634058f6139b63e] [formerly 18703581c4f27f251571b713ce27492a983d43db] [formerly c6121b17e0d6c3212383c3f8de47fa8e82bfbcc0 [formerly f87941de1d2ad239b27fa085c51252fd1c37da53] [formerly 1e2565064fef1a1ad19128eb4064aba13d402963]] [formerly f3850ea46822269d9d4e4737fa2e539758e53804 [formerly 94b1efc1bc7b0c0d6572f52e5b023dac7ee827fc]]
Former-commit-id: bd956e7f71c414ea3f643a20ad29a80dd40b1c21
Former-commit-id: e91debf4831efa4ce797113fb0d87e60ae65cabb
Former-commit-id: 04140f845b5c141aa6abba2fc7a3af8a75d808ec
Former-commit-id: d95e3824ce9f748f9fe17dd0dca8785894a4266d