Make hack in official/vision/detection models that enters the backend keras graph stop happening once we enable the Functional API KerasTensors refactoring:
As a workaround hack for the tf op layer conversion being fragile, the detection models have to explicitly enter the Keras backend graph.

When we enable the KerasTensors refactoring of the Functional API internals, the op layer conversion will be much more reliable and this hack will not be necessary. In addition, the hack actually causes the models to break when we enable the refactoring (because it causes tensors to leak out of a graph).

So, this CL changes the existing hack to stop applying once we've enabled the KerasTensors refactoring.

PiperOrigin-RevId: 322229802

Former-commit-id: 9e4d621e3152de38a93222705b795418e6318a9d [formerly 363a36cd2bff6b45efe433c766448c66da1a82ef]
Former-commit-id: 0c74b3bdee5621abf4a5d83b738d5e68802f418d [formerly 4acb75564dee892c49c9117b277f6ad15b68f6e8] [formerly cfb730db34d0a5416f8ac1471ecb233eae751358]
Former-commit-id: 5f51b48ad052cba231ffb6f672d89e451deb082e
Former-commit-id: d32266108868a114d034fe4d2850bb998fa0cb39