Make hack in official/vision/detection models that enters the backend keras graph stop happening once we enable the Functional API KerasTensors refactoring:
As a workaround hack for the tf op layer conversion being fragile, the detection models have to explicitly enter the Keras backend graph.

When we enable the KerasTensors refactoring of the Functional API internals, the op layer conversion will be much more reliable and this hack will not be necessary. In addition, the hack actually causes the models to break when we enable the refactoring (because it causes tensors to leak out of a graph).

So, this CL changes the existing hack to stop applying once we've enabled the KerasTensors refactoring.

PiperOrigin-RevId: 322229802

Former-commit-id: d41281f1710472201ad15b12fc83a42579d77e50 [formerly ae8845ce790c3864c18a53dfc40eefb29d86b040]
Former-commit-id: 77993abbd7b9a67c35b9da9b2b5e6057cdc2211f [formerly 29ed5cb55c9f069daf6135ab8111393ddc8358af] [formerly d3d6ed59fb5e9ce6082fd0e7a283a2b71d61a1f5]
Former-commit-id: 9f803f09eff2263cf935c40614f59a75ba3cb1b3
Former-commit-id: e235e401447ee378a57a855769cf6601596a33e4