Add pylint wrapper extracted from TensorFlow CI. (#9793)
TensorFlow uses a subset of pylint functionality. Script will download the latest pylint rules from the main repo at runtime.

Requires `python3.8 -m pip install pylint==2.4.4`.
This will eventually be run on-PR in this repo, but first we'll have to fix the ~900 existing violations.
Script should be usable manually for now -- call it from within a subdirectory to avoid processing the entire repo.

Former-commit-id: 3eedb03ce022e6116b50fb1841fd333b02ec2157
Former-commit-id: 2b4689abfc72078a1ea12ee1de0c8553cb5f46ba
Former-commit-id: 924d47f0eec2b81a846ed7048ae8e3c6b5c0c792