Add a new option "recreate_iterator_for_each_eval" into StandardEvaluator.
In current implementation of StandardEvaluator, the iterator of evaluate dataset will go back (actually create a new iterator) to the beginning when evaluate() is called every time.

In some case, the iterator need to go ahead forever. So add the new option "recreate_iterator_for_each_eval" to control this behavior.

In some case, the iterator creation is time consuming. For example, careting an iterator of distributed dataset for multi worker in eager mode. So it's better to make a repeat dataset instead of recreating an iterator.

If "recreate_iterator_for_each_eval" is True, then the iterator will go back; Otherwise, keep going ahead.

PiperOrigin-RevId: 364356169

Former-commit-id: 0e6f88488c2aa805181e7409162975dc1a535fa5
Former-commit-id: 06484be9adb578de87e131c993e4b41ee6312ef4
Former-commit-id: e14932f779583fb3c229c7c5e3411358a61dcb95