Add a new option "recreate_iterator_for_each_eval" into StandardEvaluator.
In current implementation of StandardEvaluator, the iterator of evaluate dataset will go back (actually create a new iterator) to the beginning when evaluate() is called every time.

In some case, the iterator need to go ahead forever. So add the new option "recreate_iterator_for_each_eval" to control this behavior.

In some case, the iterator creation is time consuming. For example, careting an iterator of distributed dataset for multi worker in eager mode. So it's better to make a repeat dataset instead of recreating an iterator.

If "recreate_iterator_for_each_eval" is True, then the iterator will go back; Otherwise, keep going ahead.

PiperOrigin-RevId: 364356169

Former-commit-id: 558ebf8e4690f2cc68b4043ca01ba3e1630e326f [formerly 7f2ee92dc037e88fdebe2915de65c5ab295e7b6e] [formerly cf6b3f71c9a4cc50f25e202f4913a5bcbe778aa7] [formerly 3573ab7606a462ccc4a08a3d8425dc414b9a9ae0] [formerly b5903dc45b1faf1a7c7c7dfe3881843625a6357e [formerly 0b7674b9e8f110bbb848cd8ef23d72ee49fc95af]]
Former-commit-id: 29bcfe32a97b882a12f0d74702a4de78167933ef
Former-commit-id: e1af181fb84e8999c5beafb12e126585d8b07f41
Former-commit-id: a90d3fca8027ad49c1c2fdf280965716d64b663f
Former-commit-id: a9df25027e95bfd08211e5b9d2f49e26f63b52e6